+++ app/src/main/java/com/sopt/now/presentation/home/user/UserViewHolder.kt
@@ -2,7 +2,6 @@ package com.sopt.now.presentation.home.user
 
 import androidx.recyclerview.widget.RecyclerView
 import com.sopt.now.databinding.ItemUserBinding
-import com.sopt.now.presentation.model.User
 
 class UserViewHolder(private val binding: ItemUserBinding) :
     RecyclerView.ViewHolder(binding.root) {
