+++ app/src/main/java/com/sopt/now/presentation/sign/SignActivity.kt
@@ -8,7 +8,7 @@ import com.sopt.now.core.util.context.snackBar
 import com.sopt.now.core.util.context.toast
 import com.sopt.now.databinding.ActivitySignBinding
 import com.sopt.now.presentation.login.LoginActivity
-import com.sopt.now.presentation.model.User
+import com.sopt.now.presentation.home.user.User
 import com.sopt.now.presentation.util.KeyStorage.ERROR_SIGN_ID
 import com.sopt.now.presentation.util.KeyStorage.ERROR_SIGN_NICKNAME
 import com.sopt.now.presentation.util.KeyStorage.ERROR_SIGN_PW
